storage path

+++ apps/ta-candle-store/src/domain/model/Candle.go
@@ -5,11 +5,11 @@ type Candle struct {
 	Symbol    string  `json:"symbol"`
 	Market    string  `json:"market"`
 	Precision string  `json:"precision"`
-	Ts        int32   `json:"ts"`
+	Ts        int64   `json:"ts"`
 	Open      float32 `json:"open"`
 	Close     float32 `json:"close"`
 	High      float32 `json:"high"`
 	Low       float32 `json:"low"`
 	Volume    float32 `json:"volume"`
-	Version   int8    `json:"version"`
+	Version   int64   `json:"version"`
 }
